+++ include/render/primitives/MeshPrimitive.h
@@ -91,7 +91,8 @@ namespace render {
     void buildLeaves();
     [[nodiscard]] bool isBuildableTriangle(int index0, int index1, int index2) const;
     [[nodiscard]] bool hasValidVertexIndex(int index) const;
-    std::shared_ptr<Primitive> buildLeaf(int index0, int index1, int index2) const;
+    std::shared_ptr<Primitive> buildLeaf(std::size_t faceIndex, int index0, int index1,
+                                         int index2) const;
     std::shared_ptr<render::Material> materialForFace(std::size_t faceIndex) const;
 
     std::shared_ptr<const core::MeshAsset> m_asset;